Refactor, renamed to new_member.handler
